Working idrac role
This role now successfully pulls ssl certificate information from an iDRAC 6
BMC using my estenrye/ansible image.

Next steps are to add logic to reset the SSL configuration when the
expected certificate authority is not present or when the expected
common name is not present.

Once those steps are implemented, then I will ned to implement logic to
reset the SSL when the not_before date is after the current date and
when the not_after date is less than x days from expiration.